Inline Dockerfile and gcloud commands.
Add Dockerfile and gcloud commands to README.

It isn't clear how to deploy the same function to all three environments, Cloud Functions, Cloud Run and Cloud Run on GKE.  What's missing is a Dockerfile and example gcloud commands.  There are links to the relevant quickstarts, but it made me think I was in the wrong place.