refactor!: use a union type for SignatureType
This commit refactors the SignatureType enum into an array of strings
declared [as const](https://github.com/Microsoft/TypeScript/pull/29510).
This allows the SignatureType type to be expressed as a union type,
which works a bit better when parsing a users provided string.

This is some simple refactoring in preparation for declarative function
signatures.

BREAKING CHANGE: exported SignatureType type is converted from an enum
to a union type