refactor: move the bin entrypoint out of index.ts (#377)
This commit moves the call to bootstrap the express server out of
index.ts. This is an important change now that we have declarative
function signatures because we expect developers to include an import
statement for the functions-framework. Before this change doing so would
have automatically started the express http server which makes is hard
to do things like write unit tests.