feat: add google CloudEvent types (#376)
This commit adds the output of the experiment/generate_cloudevents
pipeline to the FunctionsFrameworks. This adds type interfaces for all
known Google CloudEvent payloads, but it does not make any changes to
the FFs to consume them yet. Those changes will be made in a subsequent
PR.

All the types added by this PR can be deleted once the npm packages
produced by googleapis/google-cloudevents-nodejs is ready for
production usage.